+++ app/src/main/java/com/example/bravohealthpark/presentation/fragments/HomeFragment.java
@@ -15,7 +15,7 @@ import com.example.bravohealthpark.domain.user.dto.FindUserResponse;
 import com.example.bravohealthpark.infra.preferences.SharedPreferenceBase;
 import com.example.bravohealthpark.infra.preferences.UserPreferences;
 import com.example.bravohealthpark.infra.retrofit.RetrofitClient;
-import com.example.bravohealthpark.infra.retrofit.RetrofitService;
+import com.example.bravohealthpark.domain.user.UserRetrofitService;
 import com.example.bravohealthpark.presentation.activities.MedRegistrationActivity;
 
 import retrofit2.Call;
@@ -27,7 +27,7 @@ public class HomeFragment extends Fragment {
     private RelativeLayout GoCalender;
     private TextView textViewHelloUser;
     private static Call<FindUserResponse> call;
-    private static RetrofitService retrofitService;
+    private static UserRetrofitService userRetrofitService;
 
 
     @Override
@@ -67,7 +67,7 @@ public class HomeFragment extends Fragment {
     }
 
     private void initRetrofitServiceAndCall() {
-        retrofitService = RetrofitClient.getApiService(RetrofitService.class);
-        call = retrofitService.sendFindMyUserRequest();
+        userRetrofitService = RetrofitClient.getApiService(UserRetrofitService.class);
+        call = userRetrofitService.sendFindMyUserRequest();
     }
 }
